requirements.txt
openai>=1.13.3
faiss-cpu
pypdf
python-dotenv
tiktoken
tqdm
